rework maker-kit to only control logic and handlers; maker implementation details offloaded to separate example repos @ airswap-maker-kit-examples
