+++ scripts/peers/handlers.js
@@ -58,10 +58,15 @@ function indexerCall(wallet, signerSide, senderSide, callback) {
 }
 
 function peerCall(locator, method, values, validator, callback) {
-  const client = jayson.client.http(locator)
+  let client
+  if (locator.includes('https')) {
+    client = jayson.client.https(locator)
+  } else {
+    client = jayson.client.http(locator)
+  }
   client.request(method, values, function(err, error, quote) {
     if (err) {
-      callback(`\n${chalk.yellow('Connection Error')}: ${locator}`)
+      callback(`\n${chalk.yellow('Connection Error')}: ${locator} \n ${err}`)
     } else {
       if (error) {
         callback(`\n${chalk.yellow('Maker Error')}: ${error.message}\n`)
